Create models dir and move models there
