+++ src/app/components/Image/index.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Image from './index';
 import snapshotTestHelper from '../../../__test__/snapshotTestHelper';
-import { blockContainingText } from '../../../__test__/blockHelpers';
+import { blockContainingText } from '../../models/blockHelpers';
 
 describe('Image', () => {
   describe('with no data', () => {
