+++ src/app/helpers/blockHelpers.js
