+++ src/app/models/blockHelpers.js
